Moved Merger to new namespace to break import loop
