+++ yamlpath/__init__.py
@@ -1,5 +1,3 @@
 """Core YAML Path classes."""
 from yamlpath.yamlpath import YAMLPath
 from yamlpath.processor import Processor
-from yamlpath.mergerconfig import MergerConfig
-from yamlpath.merger import Merger
