+++ yamlpath/enums/__init__.py
@@ -1,10 +1,6 @@
 """Make all the YAML Path enumerations available."""
-from .anchorconflictresolutions import AnchorConflictResolutions
 from .anchormatches import AnchorMatches
-from .aohmergeopts import AoHMergeOpts
-from .arraymergeopts import ArrayMergeOpts
 from .collectoroperators import CollectorOperators
-from .hashmergeopts import HashMergeOpts
 from .includealiases import IncludeAliases
 from .pathsearchmethods import PathSearchMethods
 from .pathsegmenttypes import PathSegmentTypes
