+++ yamlpath/exceptions/__init__.py
@@ -1,3 +1,2 @@
 """Make all of the custom YAML Path exceptions available."""
-from yamlpath.exceptions.mergeexception import MergeException
 from yamlpath.exceptions.yamlpathexception import YAMLPathException
