+++ yamlpath/merger/__init__.py
@@ -1,0 +1,3 @@
+"""Core YAML Path Merger classes."""
+from .mergerconfig import MergerConfig
+from .merger import Merger
