+++ yamlpath/merger/exceptions/__init__.py
@@ -1,0 +1,2 @@
+"""Merger exceptions."""
+from .mergeexception import MergeException
